+++ build.gradle
@@ -12,7 +12,7 @@ version = '1.0-SNAPSHOT'
 
 apply plugin: 'java'
 
-sourceCompatibility = 1.8
+sourceCompatibility = 9
 
 repositories {
     mavenCentral()
@@ -21,8 +21,6 @@ repositories {
 }
 
 apply plugin: 'org.junit.platform.gradle.plugin'
-
-// also you need to add "gdju" life template to dependencies block
 
 configurations {
     testCompile.extendsFrom compileOnly
@@ -37,6 +35,7 @@ dependencies {
             // todo: insert additional dependencies here
 
             'org.apache.logging.log4j:log4j-slf4j-impl:+',
+            'javax.xml.bind:jaxb-api:+',
             'io.vavr:vavr:+'
 
     compileOnly 'org.projectlombok:lombok:+',
@@ -44,7 +43,8 @@ dependencies {
 
     testCompile 'org.junit.jupiter:junit-jupiter-params:+',
             'org.hamcrest:java-hamcrest:+',
-            'org.mockito:mockito-core:+'
+            'org.mockito:mockito-core:+',
+            'org.springframework:spring-test:+'
 
     testRuntime 'org.junit.jupiter:junit-jupiter-engine:+'
