--- src/main/java/com/epam/jl/demo/ioc/Country.java
@@ -1,6 +1,0 @@
-package com.epam.jl.demo.ioc;
-
-public interface Country {
-    String getName();
-    String getCodeName();
-}
