--- src/main/java/com/epam/jl/demo/ioc/Person.java
@@ -1,19 +1,0 @@
-package com.epam.jl.demo.ioc;
-
-public interface Person {
-    String getFirstName();
-
-    String getLastName();
-
-    Country getCountry();
-
-    int getAge();
-
-    float getHeight();
-
-    boolean isProgrammer();
-
-    boolean isBroke();
-
-    java.util.List<Contact> getContacts();
-}
