--- src/main/java/com/epam/jl/demo/ioc/SimpleCountry.java
@@ -1,9 +1,0 @@
-package com.epam.jl.demo.ioc;
-
-import lombok.Value;
-
-@Value
-public class SimpleCountry implements Country {
-    private String name;
-    private String codeName;
-}
