+++ src/main/java/lab/model/Contact.java
@@ -1,7 +1,6 @@
-package com.epam.jl.demo.ioc;
+package lab.model;
 
 public interface Contact {
     String getType();
-
     String getValue();
 }
