+++ src/main/java/lab/model/UsualPerson.java
@@ -7,10 +7,9 @@ import java.util.List;
 @Value
 public class UsualPerson implements Person {
     private int id;
-    private String name;
+    private String firstName;
+    private String lastName;
     private Country country;
     private int age;
-    private float height;
-    private boolean isProgrammer;
-    private List<String> contacts;
+    private List<Contact> contacts;
 }
