+++ src/test/java/ioc/SpringTCFAppTest.java
@@ -1,0 +1,21 @@
+package ioc;
+
+import lab.model.UsualPerson;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import static commons.Tests.getExpectedPerson;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@ExtendWith(SpringExtension.class)
+@ContextConfiguration("classpath:application-context.xml")
+class SpringTCFAppTest {
+    private UsualPerson person;
+
+    @Test
+    void testInitPerson() {
+        assertEquals(getExpectedPerson(), person);
+    }
+}
