+++ src/test/resources/application-context.xml
@@ -1,38 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:util="http://www.springframework.org/schema/util"
+       xmlns:c="http://www.springframework.org/schema/c"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:util="http://www.springframework.org/schema/util" xmlns:c="http://www.springframework.org/schema/c"
-       xsi:schemaLocation="
-        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-        http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util.xsd">
-
-    <bean id="person1" class="com.epam.jl.demo.ioc.SimplePerson"
-          c:firstName="John"
-          c:lastName="Smith"
-          c:age="35"
-          c:isProgrammer="true"
-          c:height="1.78"
-          c:broke="false"
-          c:country-ref="country1"
-          c:contacts-ref="contacts1"/>
-
-    <bean id="country1" class="com.epam.jl.demo.ioc.SimpleCountry"
-          c:name="Russia"
-          c:codeName="RU"/>
-
-    <util:list id="contacts1" value-type="com.epam.jl.demo.ioc.Contact">
-        <bean class="com.epam.jl.demo.ioc.SimpleContact" c:type="EMAIL"
-              c:value="asd@asd.ru"/>
-        <bean class="com.epam.jl.demo.ioc.SimpleContact" c:type="TELEPHONE"
-              c:value="+55 11 99999-5555"/>
-    </util:list>
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
+        http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util.xsd http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
 
     <bean id="person" class="lab.model.UsualPerson"
           c:id="1"
-          c:name="John Smith"
+          c:firstName="John"
+          c:lastName="Smith"
           c:country-ref="country"
-          c:age="35" c:height="1.78"
-          c:isProgrammer="true"
+          c:age="35"
           c:contacts-ref="contacts"/>
 
     <bean id="country" class="lab.model.SimpleCountry"
@@ -40,8 +19,11 @@
           c:name="Russia"
           c:codeName="RU"/>
 
-    <util:list id="contacts" value-type="java.lang.String">
-        <bean class="java.lang.String" c:_0="222-33-22"/>
-        <bean class="java.lang.String" c:_0="jkhafg@kjhsd.ru"/>
+    <util:list id="contacts" value-type="lab.model.Contact">
+        <bean class="lab.model.SimpleContact"
+              c:type="TELEPHONE" c:value="222-33-22"/>
+        <bean class="lab.model.SimpleContact"
+            c:type="EMAIL" c:value="jkhafg@kjhsd.ru"/>
     </util:list>
+
 </beans>
